Introduce CLASS keyword and valid pointers (fixes #73)
This change introduces a CLASS which is distinct from a RECORD.
Instances of classes can only be accessed through pointers.
This also applies to method implementations, where methods for
classes must have a self parameters which is a valid pointer.
Additionally, almost any pointer can be dclared as a VALID
pointer, which means only another VALID pointer value can be
assigned to it.